Added flag.
New flag:
- searchDOM - searches the DOM element for the configuration object
settings and adds them to the options. Every property of the
configuration object can be rewritten as ‘data-<property>’ in the HTML.
Useful for those who can’t access the script or don’t know how to use
it. If not needed, set to false to improve the speed.